requirements.txt
torch
torchvision
matplotlib
pandas
numpy
Pillow
pyyaml
tqdm
gradio